+++ test/unit/angularjs/rails/resourceProviderSpec.js
@@ -2,42 +2,42 @@ describe('resource provider factory config', function () {
     'use strict';
 
     it('should allow disabling root wrapping globally', function () {
-        module('rails', function (railsResourceFactoryProvider) {
-            expect(railsResourceFactoryProvider.enableRootWrapping(false)).toBe(railsResourceFactoryProvider);
+        module('rails', function (RailsResourceProvider) {
+            expect(RailsResourceProvider.rootWrapping(false)).toBe(RailsResourceProvider);
         });
 
         inject(function (railsResourceFactory) {
-            expect(railsResourceFactory({name: 'test', url: '/test'}).enableRootWrapping).toBe(false);
+            expect(railsResourceFactory({name: 'test', url: '/test'}).config.rootWrapping).toBe(false);
         });
     });
 
     it('should allow setting updateMethod globally', function () {
-        module('rails', function (railsResourceFactoryProvider) {
-            expect(railsResourceFactoryProvider.updateMethod('patch')).toBe(railsResourceFactoryProvider);
+        module('rails', function (RailsResourceProvider) {
+            expect(RailsResourceProvider.updateMethod('patch')).toBe(RailsResourceProvider);
         });
 
         inject(function (railsResourceFactory) {
-            expect(railsResourceFactory({name: 'test', url: '/test'}).updateMethod).toBe('patch');
+            expect(railsResourceFactory({name: 'test', url: '/test'}).config.updateMethod).toBe('patch');
         });
     });
 
     it('should allow setting http headers options globally', function () {
-        module('rails', function (railsResourceFactoryProvider) {
-            expect(railsResourceFactoryProvider.httpConfig({headers: {'test': "header"}})).toBe(railsResourceFactoryProvider);
+        module('rails', function (RailsResourceProvider) {
+            expect(RailsResourceProvider.httpConfig({headers: {'test': "header"}})).toBe(RailsResourceProvider);
         });
 
         inject(function (railsResourceFactory) {
-            expect(railsResourceFactory({name: 'test', url: '/test'}).httpConfig.headers).toEqualData({'Accept': 'application/json', 'Content-Type': 'application/json', 'test': 'header'});
+            expect(railsResourceFactory({name: 'test', url: '/test'}).config.httpConfig.headers).toEqualData({'Accept': 'application/json', 'Content-Type': 'application/json', 'test': 'header'});
         });
     });
 
     it('should allow setting default query parameters options globally', function () {
-        module('rails', function (railsResourceFactoryProvider) {
-            expect(railsResourceFactoryProvider.defaultParams({'test': "1"})).toBe(railsResourceFactoryProvider);
+        module('rails', function (RailsResourceProvider) {
+            expect(RailsResourceProvider.defaultParams({'test': "1"})).toBe(RailsResourceProvider);
         });
 
         inject(function (railsResourceFactory) {
-            expect(railsResourceFactory({name: 'test', url: '/test'}).defaultParams).toEqualData({'test': '1'});
+            expect(railsResourceFactory({name: 'test', url: '/test'}).config.defaultParams).toEqualData({'test': '1'});
         });
     });
 });