+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+<a name="1.0.0"></a>
+# 1.0.0
+## Bug Fixes
+
+## Features
+- Added <code>configure</code> to allow changing configuration options after the resource has been initially configured.
+
+## Breaking Changes
+- <code>railsResourceFactoryProvider</code> settings have been moved to <code>RailsResourceProvider</code>
+- <code>wrapData</code> config option has been renamed <code>rootWrapping</code>
+- All resource settings are now stored under the <code>config</code> property on the resource and should be modified using the <code>configure</code> function.
+- The following resource settings have been renamed:
+    - <code>enableRootWrapping</code> was renamed <code>rootWrapping</code>
+    - <code>rootName</code> was renamed <code>name</code>
+    - <code>rootPluralName</code> was renamed <code>pluralName</code>
+
 <a name="0.2.3"></a>
 # 0.2.3
 ## Bug Fixes
@@ -25,14 +41,14 @@
 - Removed default transformers and interceptors
     - railsFieldRenamingTransformer and railsFieldRenamingInterceptor have been removed completely and replaced by the serializers
     - railsRootWrappingTransformer/Interceptor are no longer configured by the transformers/interceptors configuration option and is instead
-      configured by the <code>enableRootWrapping</code> option.
+      configured by the <code>wrapData</code> option.
 - Interceptors added using <code>beforeRequest</code> are run before any deserialization so the fields names have not been camelized.
 
 ## Bug Fixes
 
 ## Features
 - Added serializers to replace old field renaming logic and to give users a lot more flexibility in customizing the (de)serialization process
-- Added <code>enableRootWrapping</code> configuration option to be able to turn off root wrapping instead
+- Added <code>rootWrapping</code> configuration option to be able to turn off root wrapping instead
 - Added path option to <code>$url</code> methods to make it easier to construct a nested url.
 
 <a name="0.1.7"></a>